+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.controller;
 
 import com.techcourse.model.User;
 import com.techcourse.service.UserService;
