+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.controller;
 
 import com.techcourse.service.UserService;
 import java.util.Map;
