+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,10 +6,10 @@ import java.net.Socket;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.controller.Controller;
 import org.apache.coyote.http11.error.ErrorMapper;
-import org.apache.coyote.http11.handler.GreetingController;
-import org.apache.coyote.http11.handler.HttpResourceController;
-import org.apache.coyote.http11.handler.LoginController;
-import org.apache.coyote.http11.handler.RegisterController;
+import com.techcourse.controller.GreetingController;
+import org.apache.coyote.http11.controller.HttpResourceController;
+import com.techcourse.controller.LoginController;
+import com.techcourse.controller.RegisterController;
 import org.apache.coyote.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
